# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estudent",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/estudent/api",
            "description": "Rest API Server"
        }
    ],
    "paths": {
...
        "/exam-registrations": {
            "get": {
                "tags": [
                    "Common Routes"
                ],
                "summary": "Retrieve exam registrations",
                "operationId": "exam-registrations/index",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExamRegistrationFiltersDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exam registrations retrieved successfully"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Common Routes"
                ],
                "summary": "Store new exam registration",
                "operationId": "exam-registrations/store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubmitExamRegistrationDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ExamRegistration stored successfully."
                    },
                    "400": {
                        "description": "Validation errors:.../CourseExam with provided course_id:x, and exam_period_id:x doesn't exist/Registration no longer in progress for given exam period"
                    },
                    "409": {
                        "description": "ExamRegistration for provided course_id:x, and exam_period_id:x and student_id:x already exist"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/exam-registrations/student": {
            "get": {
                "tags": [
                    "Common Routes"
                ],
                "summary": "Retrieve exam registrations for the authenticated student",
                "description": "Retrieve exam registrations for the authenticated student, applying filters but forcing student_id to user->id",
                "operationId": "488fd754c9c0d8e9538e4f116668bf36",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExamRegistrationFiltersDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exam registrations retrieved successfully"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/exam-registrations/not-graded": {
            "get": {
                "tags": [
                    "Common Routes"
                ],
                "summary": "Retrieve not graded exam registrations only for logged in/passed student",
                "operationId": "ab0b64e80d99320635581146113cadb7",
                "parameters": [
                    {
                        "name": "student-id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Not graded exam registrations retrieved successfully"
                    },
                    "404": {
                        "description": "Student with id student_id not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/exam-registrations/not-graded/all": {
            "get": {
                "tags": [
                    "Admin Routes"
                ],
                "summary": "Retrieve not graded exam registrations for all students",
                "operationId": "600051d4b1752207b2f8a9eca9cf7fb7",
                "responses": {
                    "200": {
                        "description": "All not graded exam registrations retrieved successfully"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/exam-registrations/{examRegistrationId}": {
            "put": {
                "tags": [
                    "Admin Routes"
                ],
                "summary": "Update existing exam registration",
                "operationId": "exam-registrations/update",
                "parameters": [
                    {
                        "name": "exam-registration-id",
                        "in": "path",
                        "description": "ID of the exam registration to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExamRegistrationUpdateDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Updated successfully"
                    },
                    "403": {
                        "description": "Error: Forbidden"
                    },
                    "404": {
                        "description": "ExamRegistration not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Common Routes"
                ],
                "summary": "Delete existing exam registration",
                "operationId": "exam-registrations/destroy",
                "parameters": [
                    {
                        "name": "exam-registration-id",
                        "in": "path",
                        "description": "ID of the exam registration to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "ExamRegistration not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ExamRegistrationFiltersDTO": {
                "properties": {
                    "search-text": {
                        "type": "string",
                        "example": ""
                    },
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "page-size": {
                        "type": "integer",
                        "example": 10
                    },
                    "include-not-graded": {
                        "type": "boolean",
                        "example": false
                    },
                    "include-failed": {
                        "type": "boolean",
                        "example": false
                    },
                    "include-passed": {
                        "type": "boolean",
                        "example": false
                    },
                    "student-id": {
                        "type": "integer",
                        "example": 123
                    }
                },
                "type": "object"
            },
            "SubmitExamRegistrationDTO": {
                "required": [
                    "course-exam-id",
                    "student-id"
                ],
                "properties": {
                    "course-exam-id": {
                        "description": "ID of the CourseExam",
                        "type": "integer",
                        "example": 12
                    },
                    "student-id": {
                        "description": "ID of the Student",
                        "type": "integer",
                        "example": 34
                    }
                },
                "type": "object"
            },
            "ExamRegistrationUpdateDTO": {
                "required": [
                    "has-attended"
                ],
                "properties": {
                    "mark": {
                        "type": "integer",
                        "example": 5
                    },
                    "has-attended": {
                        "type": "boolean",
                        "example": true
                    },
                    "comment": {
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Bearer Token",
                "in": "header",
                "scheme": "https",
                "name": "Authorization"
            }
        }
    },
...
    "security": [
        {
            "oauth2_security_example": [
                "read",
                "write"
            ],
            "passport": []
        }
    ]
}